move HierarchicalObject to dto package
